v2.2.3-beta

+++ www/class/includes/changeEnv.php
@@ -7,67 +7,79 @@ trait changeEnv {
         global $DEFAULT_ENV;
         global $LAST_ENV;
 
+        ob_start();
+
         /**
          *  1. Génération du tableau récapitulatif de l'opération
          */
-        echo "<table>
+        echo "<h3>NOUVEL ENVIRONNEMENT</h3>";
+        echo "<table class=\"op-table\">
         <tr>
-            <td>Nom du repo :</td>
+            <th>Nom du repo :</th>
             <td><b>{$this->repo->name}</b></td>
         </tr>";
         if ($OS_FAMILY == "Debian") {
             echo "<tr>
-                <td>Distribution :</td>
+                <th>Distribution :</th>
                 <td><b>{$this->repo->dist}</b></td>
             </tr>
             <tr>
-                <td>Section :</td>
+                <th>Section :</th>
 	            <td><b>{$this->repo->section}</b></td>
 	        </tr>";
         }
         echo '<tr>
-        <td>Environnement source :</td>';
-        if ($DEFAULT_ENV === $LAST_ENV) { // Cas où il n'y a qu'un seul env
-            echo "<td class=\"td-redbackground\"><span>{$this->repo->env}</span></td></tr>";
-        } elseif ($this->repo->env === $DEFAULT_ENV) { 
-            echo "<td class=\"td-whitebackground\"><span>{$this->repo->env}</span></td></tr>";
-        } elseif ($this->repo->env === $LAST_ENV) {
-            echo "<td class=\"td-redbackground\"><span>{$this->repo->env}</span></td></tr>";
-        } else {
-            echo "<td class=\"td-whitebackground\"><span>{$this->repo->env}</span></td></tr>";
-        }
+            <th>Environnement source :</th>';
+            if ($DEFAULT_ENV === $LAST_ENV) { // Cas où il n'y a qu'un seul env
+                echo "<td class=\"td-redbackground\"><span>{$this->repo->env}</span></td>";
+            } elseif ($this->repo->env === $DEFAULT_ENV) { 
+                echo "<td class=\"td-whitebackground\"><span>{$this->repo->env}</span></td>";
+            } elseif ($this->repo->env === $LAST_ENV) {
+                echo "<td class=\"td-redbackground\"><span>{$this->repo->env}</span></td>";
+            } else {
+                echo "<td class=\"td-whitebackground\"><span>{$this->repo->env}</span></td>";
+            }
+        echo '</tr>';
         echo "<tr>
-            <td>Nouvel environnement :</td>";
-        if ($DEFAULT_ENV === $LAST_ENV) { // Cas où il n'y a qu'un seul env
-            echo "<td class=\"td-redbackground\"><span>{$this->repo->newEnv}</span></td></tr>";
-        } elseif ($this->repo->newEnv === $DEFAULT_ENV) { 
-            echo "<td class=\"td-whitebackground\"><span>{$this->repo->newEnv}</span></td></tr>";
-        } elseif ($this->repo->newEnv === $LAST_ENV) {
-            echo "<td class=\"td-redbackground\"><span>{$this->repo->newEnv}</span></td></tr>";
-        } else {
-            echo "<td class=\"td-whitebackground\"><span>{$this->repo->newEnv}</span></td></tr>";
-        }
+            <th>Nouvel environnement :</th>";
+            if ($DEFAULT_ENV === $LAST_ENV) { // Cas où il n'y a qu'un seul env
+                echo "<td class=\"td-redbackground\"><span>{$this->repo->newEnv}</span></td>";
+            } elseif ($this->repo->newEnv === $DEFAULT_ENV) { 
+                echo "<td class=\"td-whitebackground\"><span>{$this->repo->newEnv}</span></td>";
+            } elseif ($this->repo->newEnv === $LAST_ENV) {
+                echo "<td class=\"td-redbackground\"><span>{$this->repo->newEnv}</span></td>";
+            } else {
+                echo "<td class=\"td-whitebackground\"><span>{$this->repo->newEnv}</span></td>";
+            }
+        echo '</tr>';
         if (!empty($this->repo->description)) {
             echo "<tr>
-            <td>Description :</td>
-            <td><b>{$this->repo->description}</b></td>
+                <th>Description :</th>
+                <td><b>{$this->repo->description}</b></td>
             </tr>";
         }
         echo '</table>';
+
+        $this->log->steplog(1);
+        $this->log->steplogInitialize('createEnv');
+        $this->log->steplogTitle("CREATION DE L'ENVIRONNEMENT ".envtag($this->repo->newEnv)."");
+        $this->log->steplogLoading();
 
         /**
          *  2. On vérifie si le repo existe
          */
         if ($OS_FAMILY == "Redhat") {
             if ($this->repo->existsEnv($this->repo->name, $this->repo->env) === false) {
-                echo '<p><span class="redtext">Erreur :</span> ce repo n\'existe pas en '.envtag($this->repo->env).'</p>';
-                return false;
+                throw new Exception('ce repo n\'existe pas en '.envtag($this->repo->env).'');
+                /*echo '<p><span class="redtext">Erreur :</span> ce repo n\'existe pas en '.envtag($this->repo->env).'</p>';
+                return false;*/
             }
         }
         if ($OS_FAMILY == "Debian") {
             if ($this->repo->section_existsEnv($this->repo->name, $this->repo->dist, $this->repo->section, $this->repo->env) === false) {
-                echo '<p><span class="redtext">Erreur :</span> cette section n\'existe pas en '.envtag($this->repo->env).'</p>';
-                return false;
+                throw new Exception('cette section n\'existe pas en '.envtag($this->repo->env).'');
+                /*echo '<p><span class="redtext">Erreur :</span> cette section n\'existe pas en '.envtag($this->repo->env).'</p>';
+                return false;*/
             }
         }
 
@@ -94,20 +106,14 @@ trait changeEnv {
         $this->repo->time = $resultTime['Time'];
         $this->repo->signed = $resultSigned['Signed'];
         $this->repo->type = $resultType['Type'];
-        if (!empty($resultGroupId['Id_group'])) {
-            $this->repo->group = $resultGroupId['Id_group'];
-        }
+        if (!empty($resultGroupId['Id_group'])) $this->repo->group = $resultGroupId['Id_group'];
 
         /**
          *  4. Si on n'a pas transmis de description, on va conserver celle actuellement en place sur $this->repo->newEnv si existe. Cependant si il n'y a pas de description ou qu'aucun repo n'existe actuellement dans l'env $this->repo->newEnv alors celle-ci restera vide
          */
         if (empty($this->repo->description)) {
-            if ($OS_FAMILY == "Redhat") {
-                $result = $this->db->querySingleRow("SELECT Description from repos WHERE Name = '{$this->repo->name}' AND Env = '{$this->repo->newEnv}' AND Status = 'active'");
-            }
-            if ($OS_FAMILY == "Debian") {
-                $result = $this->db->querySingleRow("SELECT Description from repos WHERE Name = '{$this->repo->name}' AND Dist = '{$this->repo->dist}' AND Section = '{$this->repo->section}' AND Env = '{$this->repo->newEnv}' AND Status = 'active'");
-            }
+            if ($OS_FAMILY == "Redhat") $result = $this->db->querySingleRow("SELECT Description from repos WHERE Name = '{$this->repo->name}' AND Env = '{$this->repo->newEnv}' AND Status = 'active'");
+            if ($OS_FAMILY == "Debian") $result = $this->db->querySingleRow("SELECT Description from repos WHERE Name = '{$this->repo->name}' AND Dist = '{$this->repo->dist}' AND Section = '{$this->repo->section}' AND Env = '{$this->repo->newEnv}' AND Status = 'active'");
             if (!empty($result)) {
                 $this->repo->description = $result['Description'];
             } else {
@@ -120,14 +126,16 @@ trait changeEnv {
          */
         if ($OS_FAMILY == "Redhat") {
             if ($this->repo->existsDateEnv($this->repo->name, $this->repo->date, $this->repo->newEnv) === true) {
-                echo "<p><span class=\"redtext\">Erreur :</span> ce repo est déjà en {$this->repo->newEnv} au <b>{$this->repo->dateFormatted}</b></p>";
-                return false;
+                throw new Exception("ce repo est déjà en ".envtag($this->repo->newEnv)." au <b>{$this->repo->dateFormatted}</b>");
+                /*echo "<p><span class=\"redtext\">Erreur :</span> ce repo est déjà en {$this->repo->newEnv} au <b>{$this->repo->dateFormatted}</b></p>";
+                return false;*/
             }
         }
         if ($OS_FAMILY == "Debian") {
             if ($this->repo->section_existsDateEnv($this->repo->name, $this->repo->dist, $this->repo->section, $this->repo->date, $this->repo->newEnv) === true) {
-                echo "<p><span class=\"redtext\">Erreur :</span> cette section est déjà en {$this->repo->newEnv} au <b>{$this->repo->dateFormatted}</b></p>";
-                return false;
+                throw new Exception("cette section est déjà en ".envtag($this->repo->newEnv)." au <b>{$this->repo->dateFormatted}</b>");
+                /*echo "<p><span class=\"redtext\">Erreur :</span> cette section est déjà en {$this->repo->newEnv} au <b>{$this->repo->dateFormatted}</b></p>";
+                return false;*/
             }
         }
 
@@ -147,9 +155,7 @@ trait changeEnv {
                 /**
                  *  Suppression du lien symbolique (on sait jamais si il existe)
                  */
-                if (file_exists("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}")) {
-                    unlink("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}");
-                }
+                if (file_exists("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}")) unlink("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}");
     
                 /**
                  *  Création du lien symbolique
@@ -166,6 +172,11 @@ trait changeEnv {
                  */
                 $this->repo->id = $this->db->lastInsertRowID();
 
+                /**
+                 *  Clôture de l'étape en cours
+                 */
+                $this->log->steplogOK();
+
             /**
              *  Cas 2 : Il y a déjà une version en $this->repo->newEnv qui va donc passer en archive. Modif du lien symbo + passage de la version précédente en archive :
              */
@@ -174,14 +185,17 @@ trait changeEnv {
                 /**
                  *  Suppression du lien symbolique
                  */
-                if (file_exists("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}")) {
-                    unlink("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}");
-                }
+                if (file_exists("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}")) unlink("${REPOS_DIR}/{$this->repo->name}_{$this->repo->newEnv}");
 
                 /**
                  *  Création du lien symbolique
                  */
                 exec("cd ${REPOS_DIR}/ && ln -sfn {$this->repo->dateFormatted}_{$this->repo->name}/ {$this->repo->name}_{$this->repo->newEnv}");
+
+                /**
+                 *  Clôture de l'étape en cours
+                 */
+                $this->log->steplogOK();
 
                 /**
                  *  Passage de l'ancienne version de $this->repo->newEnv en archive
@@ -192,17 +206,24 @@ trait changeEnv {
                 $old_repoDateFormatted = DateTime::createFromFormat('Y-m-d', $old_repoDate)->format('d-m-Y');
                 $result = $this->db->querySingleRow("SELECT Description FROM repos WHERE Name = '{$this->repo->name}' AND Env = '{$this->repo->newEnv}' AND Status = 'active'");
                 $old_repoDescription = $result['Description'];
+
+                /**
+                 *  Création d'un nouveau div dans le log pour l'étape d'archivage
+                 */
+                $this->log->steplog(2);
+                $this->log->steplogInitialize('archiveRepo');
+                $this->log->steplogTitle("ARCHIVAGE DU MIROIR DU $old_repoDateFormatted");
+                $this->log->steplogLoading();
                 
                 /**
                  *  Renommage du répertoire en archived_
                  *  Si un répertoire du même nom existe déjà alors on le supprime
                  */
-                if (is_dir("${REPOS_DIR}/archived_${old_repoDateFormatted}_{$this->repo->name}")) {
-                    exec("rm -rf '${REPOS_DIR}/archived_${old_repoDateFormatted}_{$this->repo->name}'");
-                }
+                if (is_dir("${REPOS_DIR}/archived_${old_repoDateFormatted}_{$this->repo->name}")) exec("rm -rf '${REPOS_DIR}/archived_${old_repoDateFormatted}_{$this->repo->name}'");
                 if (!rename("${REPOS_DIR}/${old_repoDateFormatted}_{$this->repo->name}", "${REPOS_DIR}/archived_${old_repoDateFormatted}_{$this->repo->name}")) {
-                    echo "<p><span class=\"redtext\">Erreur :</span> un problème est survenu lors du passage de l'ancienne version du $old_repoDateFormatted en archive</p>";
-                    return false;
+                    throw new Exception("un problème est survenu lors du passage de l'ancienne version du <b>$old_repoDateFormatted</b> en archive");
+                    /*echo "<p><span class=\"redtext\">Erreur :</span> un problème est survenu lors du passage de l'ancienne version du $old_repoDateFormatted en archive</p>";
+                    return false;*/
                 }
 
                 /**
@@ -221,6 +242,11 @@ trait changeEnv {
                  */
                 exec("find ${REPOS_DIR}/archived_${old_repoDateFormatted}_{$this->repo->name}/ -type f -exec chmod 0660 {} \;");
                 exec("find ${REPOS_DIR}/archived_${old_repoDateFormatted}_{$this->repo->name}/ -type d -exec chmod 0770 {} \;");
+
+                /**
+                 *  Clôture de l'étape en cours
+                 */
+                $this->log->steplogOK();
             }
         }
     
@@ -231,11 +257,9 @@ trait changeEnv {
             if ($this->repo->section_existsEnv($this->repo->name, $this->repo->dist, $this->repo->section, $this->repo->newEnv) === false) {
                 
                 /**
-                 *  Suppression du lien symbolique (on sait jamais si il existe)
+                 *  Suppression du lien symbolique (on ne sait jamais si il existe)
                  */
-                if (file_exists("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}")) {
-                    unlink("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}");
-                }
+                if (file_exists("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}")) unlink("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}");
 
                 /**
                  *  Création du lien symbolique
@@ -251,6 +275,11 @@ trait changeEnv {
                  *  Récupération de l'ID du repos précédemment inséré car on va en avoir besoin pour l'ajouter au même groupe que le repo sur $this->repo->env
                  */
                 $this->repo->id = $this->db->lastInsertRowID();
+
+                /**
+                 *  Clôture de l'étape en cours
+                 */
+                $this->log->steplogOK();
             
             /**
              *  Cas 2 : Il y a déjà une version en $this->repo->newEnv qui va donc passer en archive. Modif du lien symbo + passage de la version précédente en archive :
@@ -260,14 +289,17 @@ trait changeEnv {
                 /**
                  *  Suppression du lien symbolique
                  */
-                if (file_exists("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}")) {
-                    unlink("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}");
-                }
+                if (file_exists("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}")) unlink("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->section}_{$this->repo->newEnv}");
 
                 /**
                  *  Création du lien symbolique
                  */
                 exec("cd ${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/ && ln -sfn {$this->repo->dateFormatted}_{$this->repo->section}/ {$this->repo->section}_{$this->repo->newEnv}");
+
+                /**
+                 *  Clôture de l'étape en cours
+                 */
+                $this->log->steplogOK();
 
                 /**
                  *  Passage de l'ancienne version de $this->repo->newEnv en archive
@@ -280,15 +312,22 @@ trait changeEnv {
                 $old_repoDescription = $result['Description'];
 
                 /**
+                 *  Création d'un nouveau div dans le log pour l'étape d'archivage
+                 */
+                $this->log->steplog(2);
+                $this->log->steplogInitialize('archiveRepo');
+                $this->log->steplogTitle("ARCHIVAGE DU MIROIR DU $old_repoDateFormatted");
+                $this->log->steplogLoading();
+
+                /**
                  *  Renommage du répertoire en archived_
                  *  Si un répertoire du même nom existe déjà alors on le supprime
                  */
-                if (is_dir("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/archived_${old_repoDateFormatted}_{$this->repo->section}")) {
-                    exec("rm -rf '${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/archived_${old_repoDateFormatted}_{$this->repo->section}'");
-                }
+                if (is_dir("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/archived_${old_repoDateFormatted}_{$this->repo->section}")) exec("rm -rf '${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/archived_${old_repoDateFormatted}_{$this->repo->section}'");
                 if (!rename("${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/${old_repoDateFormatted}_{$this->repo->section}", "${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/archived_${old_repoDateFormatted}_{$this->repo->section}")) {
-                    echo "<p><span class=\"redtext\">Erreur :</span> un problème est survenu lors du passage de l'ancienne version du $old_repoDateFormatted en archive</p>";
-                    return false;
+                    throw new Exception("un problème est survenu lors du passage de l'ancienne version du <b>$old_repoDateFormatted</b> en archive");
+                    /*echo "<p><span class=\"redtext\">Erreur :</span> un problème est survenu lors du passage de l'ancienne version du $old_repoDateFormatted en archive</p>";
+                    return false;*/
                 }
 
                 /**
@@ -307,6 +346,11 @@ trait changeEnv {
                  */
                 exec("find ${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/archived_${old_repoDateFormatted}_{$this->repo->section}/ -type f -exec chmod 0660 {} \;");
                 exec("find ${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/archived_${old_repoDateFormatted}_{$this->repo->section}/ -type d -exec chmod 0770 {} \;");
+
+                /**
+                 *  Clôture de l'étape en cours
+                 */
+                $this->log->steplogOK();
             }
         }
 
@@ -330,7 +374,7 @@ trait changeEnv {
             exec("find ${REPOS_DIR}/{$this->repo->name}/{$this->repo->dist}/{$this->repo->dateFormatted}_{$this->repo->section}/ -type d -exec chmod 0770 {} \;");          
         }
 
-        echo '<p>Terminé <span class="greentext">✔</span></p>';
+        //echo '<p>Terminé <span class="greentext">✔</span></p>';
 
         $this->repo->cleanArchives();
     }
